add header for data and move there common consts
