+++ oss-internship-2020/libtiff/example/main_sandboxed.cc
@@ -12,12 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include <array>
-#include <cstdint>
 #include <cstdlib>
-#include <cstring>
 
-#include "../sandboxed.h"  // NOLINT(build/include)
+#include "../sandboxed.h"    // NOLINT(build/include)
+#include "../test/data.h"    // NOLINT(build/include)
+#include "../test/helper.h"  // NOLINT(build/include)
 #include "absl/algorithm/container.h"
 #include "absl/strings/str_join.h"
 #include "sandboxed_api/sandbox2/util/fileops.h"
@@ -26,35 +25,6 @@
 #include "tiffio.h"  // NOLINT(build/include)
 
 namespace {
-
-struct ChannelLimits {
-  uint8_t min_red;
-  uint8_t max_red;
-  uint8_t min_green;
-  uint8_t max_green;
-  uint8_t min_blue;
-  uint8_t max_blue;
-  uint8_t min_alpha;
-  uint8_t max_alpha;
-};
-
-constexpr uint32_t kRawTileNumber = 9;
-constexpr uint32_t kClusterSize = 6;
-constexpr uint32_t kChannelsInPixel = 3;
-constexpr uint32_t kTestCount = 3;
-constexpr uint32_t kImageSize = 128 * 128;
-constexpr uint32_t kClusterImageSize = 64 * 64;
-using ClusterData = std::array<uint8_t, kClusterSize>;
-
-constexpr std::array<std::pair<uint32_t, ClusterData>, kTestCount> kClusters = {
-    {{0, {0, 0, 2, 0, 138, 139}},
-     {64, {0, 0, 9, 6, 134, 119}},
-     {128, {44, 40, 63, 59, 230, 95}}}};
-
-constexpr std::array<std::pair<uint32_t, ChannelLimits>, kTestCount> kLimits = {
-    {{0, {15, 18, 0, 0, 18, 41, 255, 255}},
-     {64, {0, 0, 0, 0, 0, 2, 255, 255}},
-     {512, {5, 6, 34, 36, 182, 196, 255, 255}}}};
 
 constexpr absl::string_view kClusterErrorFormatStr =
     "Cluster %d did not match expected results.\n"
@@ -140,30 +110,6 @@ absl::Status CheckRgbaPixel(uint32_t pixel, const ChannelLimits& limits,
 }
 
 }  // namespace
-
-std::string GetFilePath(const absl::string_view dir,
-                        const absl::string_view filename) {
-  return sandbox2::file::JoinPath(dir, "test", "images", filename);
-}
-
-std::string GetFilePath(const absl::string_view filename) {
-  std::string cwd = sandbox2::file_util::fileops::GetCWD();
-  auto find = cwd.rfind("build");
-
-  std::string project_path;
-  if (find == std::string::npos) {
-    LOG(ERROR)
-        << "Something went wrong: CWD don't contain build dir. "
-        << "Please run tests from build dir or send project dir as a "
-        << "parameter: ./sandboxed /absolute/path/to/project/dir .\n"
-        << "Falling back to using current working directory as root dir.\n";
-    project_path = cwd;
-  } else {
-    project_path = cwd.substr(0, find);
-  }
-
-  return sandbox2::file::JoinPath(project_path, "test", "images", filename);
-}
 
 absl::Status LibTIFFMain(const std::string& srcfile) {
   // to use dir and file inside sapi-libtiff, use
